Refactor Google Analytics tools for traffic data retrieval
